requirements.txt
python-dotenv==1.0.0
tweepy==4.14.0
schedule==1.2.1
python-json-logger==2.0.7
